Use silent rules instead of nostdinc
A nice feature of automake is that it allows a VPATH build. That is,
a build from outside the source tree. This is disabled in uncrustify
via nostdinc in /src/Makefile.am. If the purpose of using nostdinc
is to clean up the output of the build, I recommend using
silent rules instead.