From: Todd Richmond
uncrustify needs a "join" option to go with the other lead/trail options.
The requirement is to remove extraneous newlines and merge them into the
same format they would have if newlines were not found in the original.
This is critical for function calls/deffs where arguments are not brought
back together to maximize line length

The attached diff is quite simple and resolves the test cases I tried so
far. It also undoes the splits that were added by the over-aggressive
splitting that I resolved in my previous large diff