Update src/tokenize_cleanup.cpp
Changes pertaining to bugs of >> - ID: 3555610
Resolve conflict between template and comparator followed by shift operator
The function handle_double_angle_close is called when then second '>' of  '>>' is reached.
pc is then '>' which is of comparator type. pc->prev is also of comparator type. We delete pc and make prev as CT_ARITH. We also make the next field of prev point to pc->next.  prev now behaves has the current pc.
This invokes a case of angle mismatch because earlier a '<' was opened and detected as ANGLE_OPEN, which should now be CT_COMPARE. Thus the while loop helps us detect this '<' and make it CT_COMPARE.