Bugfix: Undesired effect of disable_processing_cmt
The new option disable_processing_cmt that has been added recently had
an undesired side effect. Unfortunatly, due to its nature, it was
not revealed by the test facility.

Configuration created before the option was introduced works fine
even with versions containing the new option. But in the configuration
created with those new versions  the option was automatically set to "".
In the code this lead to unformatted text after the first comment!

Fixed by treating NULL and "" equally.