Ensure every error reported increments the error count
The error count is used to return the error state to the caller. Uncrustify has been dropping many error conditions without reporting them because many error log calls weren't matched with an increment of error count. I addressed every instance I found.

(Would be better for this to be done in the log function itself, but there are cases we do multiple err-log output in a row for a single error, and don't want to deal with avoiding redundant counts for now.)