Align statement following a label in same line
If a label is followed by a statement instead of by a newline, the
statement shall be aligned the same way as it would be without the
label, i.e. rather than

switch (i)
{
    case 0:
        foo();
j: bar();
        break;

    case 1:
        baz();
        goto j;
        break;
}

which was the result up to now, the following output will be generated:

switch (i)
{
    case 0:
        foo();
j:      bar();
        break;

    case 1:
        baz();
        goto j;
        break;
}

This closes SourceForge bug ticket #613 and GitHub issue #430.