Enable a different configuration file for the rerun test
It sometimes is necessary to use a slightly different configuration file
for the rerun test. If, for example, the configuration file defines
different input and output tab sizes and the positioning of an item
depends on the input tab size, then the output tab size of the test must
be the input tab size for the rerun test. Otherwise the rerun test may
fail.

A test number with an exclamation mark provides a specific configuration
file for the rerun test which has the same name as the original one, but
with .rerun attached to that name.