Improved functions with arguments in multiple lines
It was possible to control newlines after and before '(' and ')'
while declaring/defining functions, as well as adding newlines
after each argument. It was also possible to treat functions
with no arguments differently.

This change adds options for adding newlines when function
has some arguments, but only if they span multiple lines.
It adds options for controlling function declarations
and definitions, as well as function calls.

So if the opening '(' and closing ')' are on different lines
(after applying all other options), it is possible to add
a newline after '(', before ')', and/or after each argument.

This way it's possible to say that if function's arguments
fit in the same line then they are not broken at all,
but if they don't - each argument ends up on a separate line.
With the option to add (or not) a newline after '(' and/or before ')'.

This also fixes the behaviour when some of nl_func_*_single are specified,
but no other nl_func_* options are used (the nl_func_*_single options would
be silently ignored).