Make the "indent to token following a brace" optional
The example given in the comment for this behaviour

  { a++;
    b--; };

isn't something we ever use, but doing that indenting makes a mess of

  namespace a { namespace b {

  class SomeClass {
  }

  }} // namespace a::b

by indenting the whole contents of namespace b to match the second
"namespace" token.

The existing behaviour is preserved as the default; the new behaviour
is "indent_token_after_brace = false".