Total rewrite of C# string parsing. Because of the nesting that is possible with interpolated/verbatim/combo strings, we really need a custom loop to do it all at once.
This code should handle all cases of valid C# strings, except for unicode-isms that in general uncrustify doesn't support anyway. This new code does assume input is valid c#, and may need adjusting if it needs to be more resilient against non-compilable code.

Lots of new tests added to try various complex cases. Best we can do is force a space to be added between strings, which would cause a diff in the output. This is really a case where we need a better test framework, so that we can actually validate the tokens generated, instead of the output text. The output tokens should have exactly one STRING or STRING_MULTI for each of these top level interpolated strings.