Handle non ASCII text (emscripten interface)
Embind cant handle non ASCII chars if normal strings are used. Using wchar
is troublesome as it needs special processing.
To solve this problem the intended functions provided by emscripten are used
to convert the strings in JavaScript code.

To keep the interface as it is the functions uncrustify and loadConfig are now
loaded in JavaScript and call the now renamed _uncrustify and _loadConfig
converted C++ functions.

Input and output strings are now passed as pointer as the emscripten functions
work with those.

_loadConfig does not need to handle non ASCII chars so all of them are removed
in loadConfig before the string is passed to _loadConfig.

Uncrustify converts the JavaScript string format to UTF8 so that it can be
processed by _uncrustify.