Add option sp_after_newop_paren
Adds the option 'sp_after_newop_paren' that controls the space after a
closing parenthesis of an new operator function.

To create this option it is necessary to set the parent_type, of the
parenthesis pair that is following a new token, to CT_NEW.

The space that is in between the new token and the opening parenthesis is
already controlled by 'sp_between_new_paren'.