Fix return on empty open_memstream buffer (emscripten interface)
Problems occurred with inputs that should gereate zero length output.
A zero length check was added for this case.

As open_memstream does not initially initialize both of its parameters to
default values. When noting is written to the returned FILE ptr both
parameters buf and len remain uninitialized. To fix this both params are
now manually default initialized.