Prevent addition of newline because of vbrace
When the option code_width is set to M, a virtual close brace (vbrace) can
be placed at column N, N>M, this caused an unexpected insertion of a
newline.

To prevent this, the vbrace that is outside of the column range is checked
if it is the last chunk on the code line. If so no break (which causes
the newline) is made.