Fix #1196 unmatched extern "C" brace
A common pattern for headers that are used in both C and C++ sources
is to conditionally wrap declarations that need C linkage in an
extern C block, but guarded by preprocessor. Since uncrustify's
parser doesn't yet match braces across preprocessor blocks, the brace
close for this "extern C" was incorrectly seen as unmatched, resulting
in fatal error.

This patch suppresses this fatal error when inside a PP block, until
the parser can be improved.

Change-Id: Ie75a066326f6a6bd43c775e6097ed81210dad67f