Make `./scripts/Test_more_Options.sh` happy
While there is no support for applying the script results I find this a good thing.
We had this idea initially as well but for every test to see if we regress something
that is not obvious like flags.

Here, in p.txt, we see that on our branch we have PCF_IN_CLASS flag on that BRACE_OPEN.