Fix a couple of small bugs pointed out by gcc 7.1
The first bug has to do with the truncation of the md5_str
in backup.cpp.  snprintf was given too small of a buffer,
so it truncated the last character.  Luckily this was
the newline, so it doesn't make too much of a difference,
but this will make it more correct.

The second bug has to do with testing whether a character
is of type BRACE or PAREN.  I believe I got it right and
that we want to check if it is either a BRACE or a PAREN,
and not some logical-OR of the two.  All of the tests
pass, at least.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>