Fix invalid tests for C enums
Some of the enums in the tests are called 'not'. The output files had
that enums incorrectly formatted, because 'not' is a keyword in C++, and
uncrustify treated it as a C keyword too.

Also that enum had a syntax error, because two values weren't separated
by a comma. Fix it by putting a missing comma there.