Add special handling for CT_CASE_COLON to output_text()
In a previous commit, a `chunk_is_str(pc, ":", 1)` was removed since it
changed the behaviour of other colons.  Most likely this line was
referring to CT_CASE_COLON if e5779e3 is to be interpreted.

I cannot see an actual difference in output when using this, but I
guess it's better to be safe than sorry.