fix diff message
Setting the working directory to the script directory caused that wrong diff
paths where printed if the script was not run inside the script directory.

To fix this a perl command is used to generate a relative path from the
(initial) working directory to the script directory. This relative path is
attached to paths that are relative to the script directory.

As end result now pahts are printed that are relative to the (initial) working
directory, instead of the script directory.