fix sp_before_byref=remove on Java if statements with &&
In some languages (java, cs, ...) 'null' is handled as type.
Prior to this commit, in cases where a 'null' CT_TYPE chunk was followed by
an && operator, caused that the insufficient rvalue references check to passed
and transformed the operator into a byref.
This in turn enabled the use of sp_before_byref.

For C++ the rvalue references check works as intended because 'null' is not
handled as CT_TYPE.

To fix the issue an additional condition was added to the rvalue references
check that only passes if the  current language is C++.

Example failure from the diff file:

 public class JavaClass {
     public static void method() {
-        if (argA != null && argB != null) {
+        if (argA != null&& argB != null) {
         }
     }
 }

original commit: bc6080789982e84f5918703805dfbc152790a71d