apply fix_enum_struct_union even if there is a typdef before an enum
The fix_enum_struct_union function does set the PCF_IN_ENUM flag which is needed
for some config options related to enums.

The `sp_enum_assign = force` option had to be added to the rdan.cfg because
before this commit the enum flags where missing so that `sp_assign` was applied.
With the flags in place now `sp_assign`is not used any more and
`sp_enum_assign` is needed to recreate the expected formatting.