set flags for enum/struct if they are preceded by a typdef
Some config options related to enums and structs need to right flags to be set
to function properly.

The `sp_enum_assign = force` option had to be added to the rdan.cfg because
before this commit the enum flags where missing so that `sp_assign` was applied.
With the flags in place now `sp_assign`is not used any more and
`sp_enum_assign` is needed to recreate the expected formatting.

Some other misc refactoring of the code.