Call punc.py directly
The script to invoke punc.py was all kinds of broken (didn't fail if the
script failed, didn't use the correct path to invoke the script, didn't
use PYTHON_EXECUTABLE). Since its only purpose was to redirect the
script's stdout, which is no longer necessary as of the previous commit,
remove the script and change the custom command to invoke punc.py
directly (with the correct paths, and using PYTHON_EXECUTABLE).

Fixes #1776.