Rename sp_before_ellipsis test files
Rename the test files which are currently named 'sp_before_ellipsis' to
instead reflect the bug report for which they were created. This frees
the config names to be used for additional tests of this option that
actually follow the pattern of setting exactly the option being tested.

This also restores sp_before_ellipsis-f.cfg, per above (the old version
wasn't even setting sp_before_ellipsis!), and continues to use it for an
existing test.