Add sp_type_ellipsis
Add an option to add or remove space between a type and `...`. This
fixes the inability to suppress such space, due to this token sequence
previously being governed by sp_after_type, which treats "remove" the
same as "force" (which otherwise makes sense). This preserves the
ability to differentiate between a pack name (sp_before_ellipsis) and a
type name (now sp_type_ellipsis).

Also, remove test 60034, since the only practical thing that test was
covering is now covered by the new suite of tests added in this commit.
However, the new tests also reveal that sp_after_type is not being
applied in some instances where one would expect.