Fix parsing of `sizeof...`
Mark an ellipsis following `sizeof` with parent type CT_SIZEOF. Check
this everywhere we would otherwise check if a token is `sizeof`. This
fixes the parsing so that `sizeof(expr)` and `sizeof...(expr)` are
parsed similarly, which fixes some otherwise spectacular formatting
failures. (In particular, this fixes the final lingering failures in C++
tests 31600-31609 that were added earlier in this branch.)

Along with the aforementioned previous commits, this also fixes #1783.