Rename test directory
Rename the "output" directory to "expected", which conveys more clearly
what is contained in this directory.

Note that this breaks the entire testing framework. Said framework is
going to be completely overhauled in a follow commits; this is being
committed separately since it is a pure rename with no content changes.