Start implementing new testing framework
Create a new testing framework, based on the existing run_test.py, but
more modularized, and designed to share more logic with the CTest-driven
set of tests.

The code that's completed as of this commit should be working, but the
major remaining work is to rewrite run_tests.py to use the new
framework, and to integrate the new framework with CTest.