Use run_test.py to run tests via CTest
Modify how we set up tests with CTest to use the new run_test.py instead
of run_test.cmake. This is a sort of proof-of-concept, since we
eventually want to use run_tests.py to set up the CTest tests as well,
which will make these changes superfluous, but it demonstrates that
run_tests.py is working.