Fix tests on Windows
Tweak generation of CTest script to force use of "CMake paths". This is
a no-op on POSIX platforms, but on Windows, forces all the paths to use
'/' instead of '\' as the path separator, which is necessary to prevent
CTest from interpreting things as escape sequences. (In particular, this
is critical so that the variable substitution where the current
configuration is substituted in is expanded, rather than seen as an
escaped '$'.)