Add sanity test
Add a CTest that just runs uncrustify (with --help). This verifies that
the build has produced a non-broken executable. In particular, this will
fail in the asan build (and maybe other sanitizer builds, if we add
them) if that detects a problem. Even more particularly, it will fail as
of this commit due to a chunk of memory which is not being freed.