Treat SIZEOF as arithmetic for check_template_arg
Tweak check_template_arg to treat an argument containing a SIZEOF (which
includes `decltype`) as "arithmetic".

This fixes #1778, although it is something of a brute force approach
(but one that seems plausible given the current approach to how
check_template_arg decides whether or not to mark things as TYPE).