Fiddle with how ansicolor.py handles errors
Move the entirety of the previously existing code of ansicolor.py into a
try block, and move providing the fallback into an except block which is
now in ansicolor.py. This looks a little strange, as the entire contents
of the module is in either a try or an except, but it simplifies things
for users of the module.