Fix identification of array reference parameters
Tweak code that detects C++ casts to also detect and correctly mark
array reference parameters (rather than incorrectly marking them as C++
casts).

This could be argued (and is intended) to fix #1804, however the test is
currently failing due to undesirable application of sp_before_byref
where sp_inside_paren would be a better choice. This will be fixed in
the next commit.