Don't apply sp_before_byref after open parenthesis
Tweak application of sp_before_byref to prevent applying it between an
open parenthesis and an ampersand, which causes sp_inside_paren to be
applied instead. This makes handling of references (such as array
references or function references) where the ampersand is preceded by an
open parenthesis consistent with similar constructs that use a pointer
star (such as function pointers), which seems like the more desirable
behavior.

Combined with the previous commit, this fixes #1804.