Run whole suites, not individual tests
Add an option to control whether unit tests are registered with CTest as
individual tests, or as one test that runs the entire suite for a
language. The latter provides less granularity from CTest, but runs much
faster, especially with Python 3 (which takes shockingly long to import
modules compared to Python 2, nearly doubling the time it takes to run
the entire set of tests when run individually).
    
Running entire suites allows us to drastically reduce the per-test cost
while retaining some parallelism.