Run CTest tests in parallel
Add a small(ish) wrapper script for running CTest, that will
automatically use parallel execution unless overridden. I have tried to
make this fairly user-friendly, but really it exists for the benefit of
the CI builds, especially on Windows. While the travis builds were
already passing -j4, this will query the actual number of CPUs on the
host machine in order to generate a guess for the "best" level of
parallelism.