Improve handling of unstable tests
Always consider an unstable test an error (fixes an inconsistency
between run_tests.py and run_test.py). Use the first-pass expected
output as the input to the second pass. Allow the second-pass expected
output file to not exist (in which case the first-pass expected output
file is used instead). Remove unused / unneeded second-pass expected
output files. Correct a pointless inconsistency in c:02102.

With this configuration, the existence of a second-pass expected output
file should be considered "suspicious". For c:00453, there really isn't
a good fix, because what uncrustify is doing in that case isn't entirely
unreasonable, but the other three are genuinely fishy. However, all
tests currently "pass", which keeps CI useful.