Remove make_message
Remove make_message function, which was apparently (that is, going by
the comment) created for the sole purpose of circumventing GCC's
-Wformat. This seemingly ignoble task was (again, going by the comment)
undertaken as a work-around for compilers that did not yet support C99's
'z' modifier. Since these compilers ought to be well and truly dead at
this point, removing the function again should be feasible.