Add support for bit-flag enumerators
Add a utility class to support strongly-typed use of certain enumerators
(which have been declared as such) to be used in flag collections and
manipulated using bitwise operators, as is done in some places,
especially for iarf_e.

This implementation is rather more complete than we probably need
because I swiped it from an implementation I'd previously written.