Improve detection of variable declarations
Add a new helper function to detect variable declarations, because the
compound expression we had before was not sufficient; it failed to
detect declarations if the type was a template or used decltype. Add
a test case, and fix genuine bugs in existing test 30023.

Also add some additional logging, and remove unneeded comment (note: the
NL value is option+1 because the option is a *blank lines* option, not a
*newlines* option).

This fixes #1919.