Remove print_options
Remove print_options, which uses a somewhat quirky and different format
for describing options, compared to save_option_file. The command line
option --show-config is retained, but effectively becomes a synonym for
--update-config-with-doc.

This is useful both in eliminating a code path of unclear utility (less
maintenance; yay!), and in that `uncrustify --show-config > mine.cfg`
becomes a quick (and perhaps slightly more obvious) way to generate a
new, clean configuration file.