Create accessors for options
Create namespaced accessor functions for obtaining the value of an
option. This will simplify the code slightly, and is also notionally
type safe.

Really, this just hides the union access, and in a lot of ways is just
all around cruddy. However, the longer term goal is to replace the great
big array of unions with actual, individual class objects that WILL be
properly type safe. The accessor functions are intended as a temporary,
intermediate step along this path that mimics the eventual API. The
switch from direct use of the array of unions to these accessors will
allow us to get the main code churn out of the way, so that when we're
ready to switch to the option objects, the interesting changes won't be
buried under a mountain of noise.