Use uncrustify namespace
Add `using namespace uncrustify` to a bunch of sources in preparation
for using the new option accessors. This will permit the commit that
actually starts using them to be a purely mechanical transformation.
(Basically, this is being done first so that the changes won't be mixed
with mechanical changes, and so that the commit that starts using the
accessors will actually compile.)