Refactor options_for_QT.cpp
Add a helper class to save/override and restore options. Create an array
of instances of this, and iterate over that rather than listing out each
option every time. This allows us to name the options to be overridden
exactly once, rather than five times as in the old code.