Add target to run uncrustify on its own sources
Add a custom target (not built by default) to run uncrustify on its own
sources. Somewhat like update_cli_options, this "assumes" you have a git
repository and clobbers the sources in-place (which, for developers, is
usually desired).

This makes it much easier to check/enforce code style while hacking on
uncrustify. It is also slightly faster, since it is set up as one main
target that depends on a sub-target per source, which allows each source
to be formatted in parallel.