Ensure outputdirectory exists
Tweak py_gen to ensure that the directory for the expected output file
exists. (We could do this in the script, but the reason behaind py_gen
is that I anticipate having more uses for it in the future, using
different scripts. Doing it in the script would imply *each* script
needing to do so, whereas this way we do it in one place.)