Normalize debug output in test_cli_options.py
Modify test_cli_options.py to tweak uncrustify's output in order to
normalize some differences between debug and release builds. This allows
the script to be used with debug builds, and allows us to get rid of all
the logic to only run it against release builds.

Besides simplifying things, which is always good, this will allow the
script to be run on coverage builds, which will help with coverage.