Fix {EXPR|STMT}_START after braced initializer
When a braced initializer list is found, clear PCF_STMT_START and
PCF_EXPR_START from the following token, as these flags are not correct
in that context. (They get added erroneously to a token following a
brace, in brace_cleanup.cpp:651, because at that time we do not know
that the brace is an initializer list and not scope.)

This fixes at least misidentification of a '&' as CT_ADDR rather than
CT_ARITH in e.g. 'int{5} & 1'.