Modernize option declarations
Start converting options to use a more "modern C++" implementation.
Specifically, make each option a C++ class to hold all of the
information about the option, and to use static initialization to set
the default value. These are strongly typed (no accessing the wrong
member of a union) and all of the information is set up in exactly one
(user editable) place.

Right now, this is ONLY the new declarations, and supporting data types;
much of the actual implementation remains to be done. The actual
initialization will be done by code generated from the declarations
using a Python script, as will conversion of the values between text and
their respective enumerations. As such, THIS COMMIT DOES NOT BUILD.