Add make_options.py
Create the script to generate the option instance declarations and
registration function from options.h. Refactor some of the options
machinery to work with the new options types, and (for now) comment out
the rest. This gets us to the point where option.cpp and options.cpp
will actually build, although the rest of the code has lots of issues
still.