Change add_keyword to take strings
Change add_keyword to take the tag argument as a std::string rather than
a char*. Since the tag was immediately converted to a std::string
anyway, this is more efficient if the caller already has a std::string.
(The need to pass a char* for logging is a wash, even aside from it
being somewhat pointless to care about logging performance. We need the
std::string anyway, and the cost of calling c_str() on the std::string
is negligible.)