Don't mark extern prototype as CT_FUNC_CTOR_VAR
Determining if certain C++ expressions are function prototypes or
variable initializations is, as our comment notes, hard, and since we
aren't a C++ compiler able to tell if an identifier might be a type name
(for non-built-in types), we sometimes get it wrong. However, a
statement that starts with 'extern' can't be an initializer, so add a
check for this.