Fix newlines_func_pre_blank_lines template parsing
Fix a bug in newlines_func_pre_blank_lines that caused it to not
correctly parse template return types. Specifically, when it encountered
a `>`, it was scanning backwards to find `template`, which is incorrect
for non-template functions with template return types (and potentially
caused a crash because we then tried to dereference a non-existing
chunk). Instead, jump back to the matching `<`, and further tweak the
logic to consume a preceding `template` as part of the prototype.