Fix grouping of CT_FUNC_CLASS_PROTO
Fix incorrect check for consecutive ctor/dtor prototypes. The test
failed because it expects the next token to have a parent type
CT_FUNC_CLASS_PROTO, which is correct for regular prototypes where the
next token is the return type or a qualifier (and might be correct if
the ctor/dtor is inline or virtual or the like), but fails when the next
token is the ctor name, since in that instance, the token itself has
type CT_FUNC_CLASS_PROTO (and parent type CT_NONE).

The actual fix is trivial; just check the token's direct type as well as
its parent type against CT_FUNC_CLASS_PROTO.