Remove run_tests.{sh,bat}
Remove shell scripts to run run_tests.py. All these did was stick
'python' in front of the name of the test script, which does not really
seem valuable enough to warrant the scripts. Besides, the preferred way
to run the tests these days (at least for the sorts of non-developers
that would be likely to want scripts like these in the first place) is
to use CTest.