Make option enums strongly typed, namespaced
Refactor enumeration types used for options to be "strongly typed", and
to live in the uncrustify namespace, which we are slowly starting to
use. For convenience, also add aliases so that the qualified names don't
need to be used.

Also, convert applicable uses of iarf_e and tokenpos_e to using the new
'strong flag type' utility from the previous commit.