Add OT_BOOL → OT_IARF "upgrade" path
Add additional aliases for IARF options in order to support a transition
path for converting OT_BOOL options to OT_IARF in a way that should
minimize breakage of existing user configuration files.

This replaces the TFI option type that was recently removed, which
existed for much the same purpose.