Fix configure with CMake < 3.x (#2085)
Fiddle with how we specify an option's documentation to pacify older
versions of CMake. (Apparently, CMake ≥ 3.x supports a greater range of
escape sequences than CMake < 3.x; 2.8.12 didn't like the clever way I
was specifying a documentation string that needed to span multiple lines
while trying to keep the indentation less ugly. Fortunately, the string
naturally breaks at a ';', which allows a reasonably nice way to build
the documentation string in a separate set() to be passed to option().
While I have not attempted to determine the exact CMake version cutoff,
I suspect it is 3.0 due to introduction of raw string literals as of
that version.)