Add nl_before_class newline before comments
Fix a bug in nl_before_class where it would add the newline in between
the class and a preceding comment, iff the class (and comment) are in a
namespace. This is almost certainly not the "correct" behavior,
especially as it differs from the behavior for classes in the global
namespace.