Add katepart highlighting file
Add a skeleton XML for katepart highlighting. Add a script to generate a
"complete" XML given the template. (The script generates several keyword
lists from our sources, so that the XML does not need to be updated by
hand.) Add a target (not in ALL) to generate the "complete" XML.