Also fix newline_def_blk for one-off extern
In much the same vein as the previous commit, also fix newline_def_blk
to back up before an associated one-off extern specification. This
prevents a case where a function prototype, such as in the unit test
added by the previous commit, is affected (in a screwy place, to boot)
by nl_var_def_blk_start.