respect --no-backup
Trust the user when it asks for no backup files,
especially when it also configures input and output files,
even if that would mean that it will replace the input file
without having a backup file.  The `--no-backup` option is
recommended to be used under VCS but in any case, the user
should know what he is doing.  Addresses issue #2174.

Signed-off-by: Mihai Popescu <mihaipopescu@users.noreply.github.com>